@@ -1,57 +1,38 @@
 // frontend/src/App.js
 import React from 'react';
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link
-} from 'react-router-dom';
-import AlarmWorkbench from './components/AlarmWorkbench';
-import LogLedger from './components/LogLedger';
-import DispatchTicketManagement from './components/DispatchTicketManagement';
+import { Routes, Route } from 'react-router-dom';
+import MainLayout from './components/MainLayout';
+import Home from './pages/Home';
+import AlarmRecordList from './pages/Alarm/AlarmRecordList';
+import AlarmDetail from './pages/Alarm/AlarmDetail';
+import DispatchManagement from './pages/Dispatch/DispatchManagement';
+import HandlingManagement from './pages/Handling/HandlingManagement';
+import ArchivingManagement from './pages/Archiving/ArchivingManagement';
+import StatisticsDashboard from './pages/Statistics/StatisticsDashboard';
+import SystemSettings from './pages/Settings/SystemSettings';
+import UserManagement from './pages/User/UserManagement';
+import RoleManagement from './pages/Role/RoleManagement';
+import DepartmentManagement from './pages/Department/DepartmentManagement';
+import LogManagement from './pages/Log/LogManagement';
 
 function App() {
   return (
-    <Router>
-      <div className="App">
-        <header className="App-header">
-          <p>接报警服务管理系统 - 前端</p>
-        </header>
-        <nav>
-          <ul>
-            <li>
-              <Link to="/">首页</Link>
-            </li>
-            <li>
-              <Link to="/workbench">接警工作台</Link>
-            </li>
-            <li>
-              <Link to="/logs">日志台账</Link>
-            </li>
-            <li>
-              <Link to="/dispatch">派警单管理</Link>
-            </li>
-          </ul>
-        </nav>
-
-        <div className="page-content">
-          <Switch>
-            <Route path="/workbench">
-              <AlarmWorkbench />
-            </Route>
-            <Route path="/logs">
-              <LogLedger />
-            </Route>
-            <Route path="/dispatch">
-              <DispatchTicketManagement />
-            </Route>
-            <Route path="/">
-              <div>欢迎使用接报警服务管理系统。请从上方导航栏选择一个模块。</div>
-            </Route>
-          </Switch>
-        </div>
-      </div>
-    </Router>
+    <Routes>
+      <Route path="/" element={<MainLayout />}>
+        <Route index element={<Home />} />
+        <Route path="alarm/records" element={<AlarmRecordList />} />
+        <Route path="alarm/records/:id" element={<AlarmDetail />} />
+        <Route path="alarm/dispatch" element={<DispatchManagement />} />
+        <Route path="alarm/handling" element={<HandlingManagement />} />
+        <Route path="alarm/archiving" element={<ArchivingManagement />} />
+        <Route path="statistics" element={<StatisticsDashboard />} />
+        <Route path="settings" element={<SystemSettings />} />
+        <Route path="users" element={<UserManagement />} />
+        <Route path="roles" element={<RoleManagement />} />
+        <Route path="departments" element={<DepartmentManagement />} />
+        <Route path="logs" element={<LogManagement />} />
+      </Route>
+    </Routes>
   );
 }
 
